WIP: replacement FullTextDocument API
All points of interest are marked with `todo` macros.
