+++ Cargo.toml
@@ -30,9 +30,6 @@ version = "0.1.0"
 
 [dependencies]
 lsp-server = "~0.5.0"
-lsp-text-document = "~0.2.0"
-# NOTE: You'll get compiler errors if this version number is different than
-#       the version used by `lsp-text-document`.
 lsp-types = "~0.84.0"
 serde = "~1.0.0"
 serde_json = "~1.0.0"
